pyproject.toml
[project]
name = "agentify-medagentbench"
version = "0.1.0"
description = "MedAgentBench evaluation framework using A2A protocol and MCP servers."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "httpx>=0.27.0",
    "litellm>=1.53.8",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
    "fastapi>=0.115.12",
    "pydantic>=2.10.5",
    "requests>=2.32.3",
]
